requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-multipart==0.0.22
requests
openai-whisper==20250625
groq==1.0.0
python-dotenv==1.2.1
streamlit==1.53.1
torch==2.10.0
